Factor out common html network page
