+++ lib/ui/src/Cardano/Wallet/UI/Personal/Server.hs
@@ -69,6 +69,9 @@ import Cardano.Wallet.UI.Common.Html.Pages.Lib
     ( alertH
     , rogerH
     )
+import Cardano.Wallet.UI.Common.Html.Pages.Network
+    ( networkInfoH
+    )
 import Cardano.Wallet.UI.Common.Layer
     ( SessionLayer (..)
     , UILayer (..)
@@ -99,9 +102,6 @@ import Cardano.Wallet.UI.Personal.Handlers.Wallets
     )
 import Cardano.Wallet.UI.Personal.Html.Pages.Addresses
     ( addressesH
-    )
-import Cardano.Wallet.UI.Personal.Html.Pages.Network
-    ( networkInfoH
     )
 import Cardano.Wallet.UI.Personal.Html.Pages.Page
     ( Page (..)
